pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "LontMKM"
version = "0.1.0"
description = "Microkinetic modeling with Marcus, Eyring, and TAS rate constants"
authors = [
  { name = "Miftahussurur Hamidi Putra", email = "miftahussurur@hamidiputra.de" }
]
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "numpy",
    "scipy",
    "pyyaml",
    "matplotlib"
]
requires-python = ">=3.8"

[project.scripts]
lontmkm = "LontMKM.cli:main"

